Er release

+++ app/src/main/java/com/example/currencyex/model/Query.java
@@ -4,7 +4,7 @@ public class Query {
 
     private String from;
     private String to;
-    private Integer amount;
+    private float amount;
 
     public String getFrom() {
         return from;
@@ -22,7 +22,7 @@ public class Query {
         this.to = to;
     }
 
-    public Integer getAmount() {
+    public float getAmount() {
         return amount;
     }
